pyproject.toml
[tool.poetry]
name = "testtaskemploees"
version = "0.1.0"
description = ""
authors = ["KrivovDE <55reigen55@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"

[tool.poetry.dev-dependencies]
djangorestframework = "^3.13.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
